rename styled-components files to style.js
