--- src/components/MovieList/MovieListItem/movie-list-item.css
@@ -1,5 +1,0 @@
-.img {
-    margin-left: 10px;
-    margin-bottom: 15px;
-    height: 230px;
-}
