+++ src/components/MovieList/MovieListItem/style.js
