+++ src/components/Search/index.js
@@ -1,1 +1,1 @@
-export { default } from './Search.jsx';
+export { default } from './Search';
