+++ src/components/Search/style.js
